+++ fdbrpc/CMakeLists.txt
@@ -64,6 +64,7 @@ if(${COROUTINE_IMPL} STREQUAL libcoro)
   endif()
 endif()
 
+target_include_directories(fdbrpc PRIVATE ${CMAKE_CURRENT_SOURCE_DIR})
 target_include_directories(fdbrpc PUBLIC "${CMAKE_CURRENT_SOURCE_DIR}/include" "${CMAKE_CURRENT_BINARY_DIR}/include" PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/libeio ${Protobuf_INCLUDE_DIRS} ${gRPC_INCLUDE_DIRS})
 target_link_libraries(fdbrpc PUBLIC proto_fdbrpc_test)
 target_link_libraries(fdbrpc PUBLIC flow libb64 md5 PRIVATE rapidjson gRPC::grpc++)
