+++ fdbrpc/FlowGrpcTests.cpp
@@ -1,0 +1,53 @@
+
+/**
+ * FlowGrpcTests.actor.cpp
+ *
+ * This source file is part of the FoundationDB open source project
+ *
+ * Copyright 2013-2024 Apple Inc. and the FoundationDB project authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+#include "flow/UnitTest.h"
+#include "fdbrpc/FlowGrpc.h"
+#include "FlowGrpcTests.h"
+
+// So that tests are not optimized out. :/
+void forceLinkGrpcTests2() {}
+
+namespace fdbrpc_test {
+namespace asio = boost::asio;
+
+TEST_CASE("/fdbrpc/grpc/basic_async_client_with_coroutines") {
+	NetworkAddress addr(NetworkAddress::parse("127.0.0.1:50500"));
+	GrpcServer server(addr);
+	shared_ptr<TestEchoServiceImpl> service(make_shared<TestEchoServiceImpl>());
+	server.registerService(service);
+	Future<Void> _ = server.run();
+
+	shared_ptr<asio::thread_pool> pool = make_shared<asio::thread_pool>(4);
+	AsyncGrpcClient<TestEchoService> client(addr.toString(), pool);
+
+	try {
+		EchoRequest request;
+		request.set_message("Ping!");
+		EchoResponse response = co_await client.call(&TestEchoService::Stub::Echo, request);
+		std::cout << "Echo received: " << response.message() << std::endl;
+		ASSERT_EQ(response.message(), "Echo: Ping!");
+	} catch (Error& e) {
+		ASSERT_EQ(e.code(), error_code_grpc_error);
+		ASSERT(false);
+	}
+}
+
+}
