+++ fdbrpc/include/fdbrpc/fdbrpc.h
@@ -29,7 +29,6 @@
 #include "fdbrpc/FailureMonitor.h"
 #include "fdbrpc/networksender.actor.h"
 #include "fdbrpc/simulator.h"
-
 #ifdef WITH_SWIFT
 #include <swift/bridging>
 #endif /* WITH_SWIFT */
