+++ fdbrpc/tests/echo.proto
@@ -4,6 +4,7 @@ package fdbrpc.test;
 
 service TestEchoService {
     rpc Echo(EchoRequest) returns (EchoResponse);
+    rpc EchoRepeat10(EchoRequest) returns (stream EchoResponse);
 }
 
 message EchoRequest {
