+++ fdbserver/workloads/UnitTests.actor.cpp
@@ -56,6 +56,7 @@ void forceLinkSimKmsVaultTests();
 void forceLinkRESTSimKmsVaultTest();
 void forceLinkActorFuzzUnitTests();
 void forceLinkGrpcTests();
+void forceLinkGrpcTests2();
 
 struct UnitTestWorkload : TestWorkload {
 	static constexpr auto NAME = "UnitTests";
@@ -127,6 +128,7 @@ struct UnitTestWorkload : TestWorkload {
 		forceLinkRESTSimKmsVaultTest();
 		forceLinkActorFuzzUnitTests();
 		forceLinkGrpcTests();
+		forceLinkGrpcTests2();
 	}
 
 	Future<Void> setup(Database const& cx) override {
